separation game et menu
separation de
game: le jeu
menus: les menus
creation d'un lanceur "main"
fusion des fichier pyxres dans res2.pyxres
